As suggested by PK: move smooth_to_fwhm to volume_operations
